requirements/base.txt
# For Heroku Deploymnet
# django-heroku

# pip>=20.1
# wheel>=0.34.2
# setuptools>=40.8.0

Django>=3.1

django-allauth>=0.42.0          # Authentication, registration, account management as well as 3rd party (social) account authentication
django-crispy-forms>=1.7.2      # For Django forms
django-dbbackup>=3.2.0          # Backup and restore Django
django-environ>=0.4.5           # django-environ allows you to use Twelve-factor methodology, used by base.py
django-compressor>=2.4          # Compress CSS and JS, For Linux: sudo apt install python3-dev; $ pip install rcssmin --install-option="--without-c-extensions"
django-libsass>=0.8             # For compile scss to css
django-simple-captcha>=0.5.13   # Captcha for sign up and sign in 

# django-q>=1.3.3                 # Django task queue (Send Email), use local database as broker, Celery does't support local database as broker

hashids>=1.3.1                  # For generate encrypt id for path
pillow>=7.2.0                   # Handle image
requests>=2.24.0                # HTTP requests